Refactor migration.sql: Update data types, add new columns, and modify constraints for multiple tables

+++ packages/prisma/src/migrations/20241006164140_add_order_relative_price/migration.sql
@@ -1,2 +1,3 @@
 -- AlterTable
-ALTER TABLE "Order" ADD COLUMN "relativePrice" REAL;
+ALTER TABLE "Order"
+  ADD COLUMN "relativePrice" DOUBLE PRECISION;
